# Api/bin/Release/net8.0/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Api, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null",
    "version": "1.0"
  },
  "paths": {
...
    "/api/pos/v0.1/Stock/addition/{trnNo}": {
      "get": {
        "tags": [
          "Stock"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "trnNo",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "delete": {
        "tags": [
          "Stock"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "trnNo",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/pos/v0.1/Stock/addition": {
      "post": {
        "tags": [
          "Stock"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/StockAddPostDTO"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "put": {
        "tags": [
          "Stock"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/StockAddUpdateDTO"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/pos/v0.1/Stock/deduction/{trnNo}": {
      "get": {
        "tags": [
          "Stock"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "trnNo",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "delete": {
        "tags": [
          "Stock"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "trnNo",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/pos/v0.1/Stock/deduction": {
      "post": {
        "tags": [
          "Stock"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/StockDeductionPostDTO"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "put": {
        "tags": [
          "Stock"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/StockDeductionUpdateDTO"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    }
  },
  "definitions": {
...
    "StockAddDetailInput": {
      "type": "object",
      "properties": {
        "itemKey": {
          "format": "int32",
          "type": "integer"
        },
        "quantity": {
          "format": "double",
          "type": "number"
        },
        "costPrice": {
          "format": "double",
          "type": "number"
        },
        "salePrice": {
          "format": "double",
          "type": "number"
        },
        "unitKey": {
          "format": "int32",
          "type": "integer"
        }
      },
      "additionalProperties": false
    },
    "StockAddPostDTO": {
      "type": "object",
      "properties": {
        "trnNo": {
          "format": "int32",
          "type": "integer"
        },
        "trnDate": {
          "format": "date-time",
          "type": "string"
        },
        "locKey": {
          "format": "int32",
          "type": "integer"
        },
        "description": {
          "type": "string"
        },
        "items": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/StockAddDetailInput"
          }
        }
      },
      "additionalProperties": false
    },
    "StockAddUpdateDTO": {
      "type": "object",
      "properties": {
        "trnNo": {
          "format": "int32",
          "type": "integer"
        },
        "trnDate": {
          "format": "date-time",
          "type": "string"
        },
        "locKey": {
          "format": "int32",
          "type": "integer"
        },
        "description": {
          "type": "string"
        },
        "items": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/StockAddDetailInput"
          }
        }
      },
      "additionalProperties": false
    },
...
    "StockDeductionItemDTO": {
      "type": "object",
      "properties": {
        "itemKey": {
          "format": "int32",
          "type": "integer"
        },
        "quantity": {
          "format": "double",
          "type": "number"
        },
        "costPrice": {
          "format": "double",
          "type": "number"
        },
        "salePrice": {
          "format": "double",
          "type": "number"
        }
      },
      "additionalProperties": false
    },
    "StockDeductionPostDTO": {
      "type": "object",
      "properties": {
        "trnDate": {
          "format": "date-time",
          "type": "string"
        },
        "locKey": {
          "format": "int32",
          "type": "integer"
        },
        "description": {
          "type": "string"
        },
        "items": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/StockDeductionItemDTO"
          }
        }
      },
      "additionalProperties": false
    },
    "StockDeductionUpdateDTO": {
      "type": "object",
      "properties": {
        "trnNo": {
          "format": "int32",
          "type": "integer"
        },
        "trnDate": {
          "format": "date-time",
          "type": "string"
        },
        "locKey": {
          "format": "int32",
          "type": "integer"
        },
        "description": {
          "type": "string"
        },
        "items": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/StockDeductionItemDTO"
          }
        }
      },
      "additionalProperties": false
    },
...
  }
}